feat(Exposed systems): show relevant OS versions
VULN-2391

This makes the table with the exposed systems render the OS filter with
values only available for the current account.
