+++ src/Components/PresentationalComponents/Filters/PrimaryToolbarFilters/OsVersionFilter.js
@@ -1,69 +1,42 @@
-import React from 'react';
-import { RHEL_VERSIONS } from '../../../../Helpers/constants';
-import { conditionalFilterType } from '@redhat-cloud-services/frontend-components/ConditionalFilter';
-import { Select, SelectOption } from '@patternfly/react-core';
+import { useSelector } from 'react-redux';
+import { useLoadModule } from '@scalprum/react-core';
 import { intl } from '../../../../Utilities/IntlProvider';
 import messages from '../../../../Messages';
 
-const useOsVersionFilter = (filterValue, apply) => {
-    const [isOpen, setOpen] = React.useState(false);
-    const [numOptions, setNumOptions] = React.useState(10);
-    const versionFromNewestToOldest = RHEL_VERSIONS.slice().reverse();
+const useOsVersionFilter = (appliedVersions, apply) => {
+    /* operatingSystems are obtained from the Inventory store */
+    const operatingSystems = useSelector(({ entities }) => entities?.operatingSystems) || [];
+    const [{ toGroupSelectionValue, buildOSFilterConfig } = {}] = useLoadModule({
+        appName: 'inventory',
+        scope: 'inventory',
+        module: './OsFilterHelpers'
+    });
 
-    let selectedVersionsArray = filterValue ? filterValue.split(',') : [];
-
-    const onOptionSelect = (_, changedLabel) => {
-        const changedValue = RHEL_VERSIONS.find(version => version.label === changedLabel).value;
-        let newValue;
-
-        if (selectedVersionsArray.includes(changedValue)) {
-            newValue = selectedVersionsArray;
-            newValue.splice(newValue.indexOf(changedValue), 1);
-        }
-        else {
-            newValue = [...selectedVersionsArray, changedValue];
-        }
-
-        apply({ rhel_version: newValue.join(','), page: 1 });
-    };
-
-    const onToggle = (isOpen) => {
-        setOpen(isOpen);
-    };
-
-    const onViewMoreClick = () => {
-        setNumOptions(versionFromNewestToOldest.length);
-    };
-
-    return {
-        type: conditionalFilterType.custom,
-        label: intl.formatMessage(messages.osFilterLabel),
-        filterValues: {
-            children: (
-                <Select
-                    variant="checkbox"
-                    onToggle={onToggle}
-                    onSelect={onOptionSelect}
-                    selections={selectedVersionsArray.map(item => RHEL_VERSIONS.find(version => version.value === item).label)}
-                    isOpen={isOpen}
-                    aria-label="os-version-filter"
-                    placeholderText={intl.formatMessage(messages.osFilterPlaceholder)}
-                    {...(numOptions < versionFromNewestToOldest.length
-                        && { loadingVariant: { text: intl.formatMessage(messages.viewMore), onClick: onViewMoreClick } })}
-                    style={{ maxHeight: '420px', overflow: 'auto' }}
-                >
-                    {versionFromNewestToOldest.slice(0, numOptions).map((option, index) => (
-                        <SelectOption
-                            isDisabled={option.disabled}
-                            key={index}
-                            value={option.label}
-                            description={option?.description}
-                        />
-                    ))}
-                </Select>
+    return buildOSFilterConfig
+        ? [
+            buildOSFilterConfig(
+                {
+                    label: intl.formatMessage(messages.osFilterLabel),
+                    type: 'checkbox',
+                    id: 'rhel_version',
+                    value: toGroupSelectionValue(
+                        appliedVersions
+                            ? appliedVersions.split(',')
+                            : []
+                    ),
+                    onChange: (event, value) => {
+                    /* `versions` must be of type string, e.g., "8.9,9.0" */
+                        const versions = Object.values(value)
+                            .flatMap((versions) => Object.keys(versions))
+                            .toString();
+                        apply({ rhel_version: versions, page: 1 });
+                    },
+                    placeholder: intl.formatMessage(messages.osFilterPlaceholder)
+                },
+                operatingSystems
             )
-        }
-    };
+        ]
+        : [];
 };
 
 export default useOsVersionFilter;
