+++ src/Components/SmartComponents/SystemsExposedTable/SystemsExposedTable.js
@@ -129,6 +129,7 @@ const SystemsExposedTable = ({
             .map(column => ({ ...inventoryColumns.find(({ key }) => column.key === key), ...column }));
     };
 
+    // TODO: let InventoryTable render its own toolbar instead of using custom one
     return (
         <Stack hasGutter>
             <StackItem>
