+++ src/Components/SmartComponents/SystemsExposedTable/SystemsExposedTableToolbar.js
@@ -69,7 +69,6 @@ export const SystemsExposedTableToolbar = ({
         apply
     );
 
-    // TODO: use dynamic OS versions
     const osVersionFilter = useOsVersionFilter(
         parameters.rhel_version,
         apply
@@ -130,7 +129,7 @@ export const SystemsExposedTableToolbar = ({
                     ),
                     statusFilter(apply, parameters),
                     advisoryFilter,
-                    osVersionFilter,
+                    ...osVersionFilter,
                     remediationFilter(apply, parameters)
                 ]
             }}
