+++ src/Components/SmartComponents/SystemsPage/SystemsPage.js
@@ -101,6 +101,7 @@ const SystemsPage = () => {
     const [columnCounter, setColumnCount] = useState(0);
     useEffect(() => setColumnCount(columnCounter + 1), [columns]);
 
+    // TODO: let InventoryTable render its own toolbar instead of using custom one
     return (
         isLoading ? <Spinner centered/> :
             canReadVulnerabilityResults ? <Fragment>
