+++ src/Components/SmartComponents/SystemsPage/SystemsTableToolbar.js
@@ -2,7 +2,7 @@ import React, { Fragment, useState, useMemo } from 'react';
 import propTypes from 'prop-types';
 import { injectIntl } from 'react-intl';
 import messages from '../../../Messages';
-import { useDispatch, useSelector } from 'react-redux';
+import { useDispatch } from 'react-redux';
 import { dataShape } from '../../../Helpers/MiscHelper';
 import { fetchSystems } from '../../../Store/Actions/Actions';
 import { PrimaryToolbar } from '@redhat-cloud-services/frontend-components/PrimaryToolbar';
@@ -18,7 +18,7 @@ import {
     addNotification,
     clearNotifications
 } from '@redhat-cloud-services/frontend-components-notifications/redux';
-import { useLoadModule } from '@scalprum/react-core';
+import useOsVersionFilter from '../../PresentationalComponents/Filters/PrimaryToolbarFilters/OsVersionFilter';
 
 const SystemsTableToolbar = ({
     selectedRows,
@@ -34,13 +34,6 @@ const SystemsTableToolbar = ({
     const [exportPDF, setExportPDF] = useState(false);
     const { apply, handleSelect, doOptOut, setColumnModalOpen } = methods;
     const dispatch = useDispatch();
-    /* operatingSystems are obtained from the Inventory store */
-    const operatingSystems = useSelector(({ entities }) => entities?.operatingSystems) || [];
-    const [{ toGroupSelectionValue, buildOSFilterConfig } = {}] = useLoadModule({
-        appName: 'inventory',
-        scope: 'inventory',
-        module: './OsFilterHelpers'
-    });
 
     const downloadReport = format => {
         let params = { ...parameters };
@@ -90,6 +83,11 @@ const SystemsTableToolbar = ({
         fetchResource: ops => fetchSystems({ ...parameters, ...ops })
     });
 
+    const osVersionFilter = useOsVersionFilter(
+        parameters.rhel_version,
+        apply
+    );
+
     const filterConfigItems = [
         useSearchFilter(
             'filter',
@@ -99,30 +97,7 @@ const SystemsTableToolbar = ({
             apply
         ),
         ...(canReadExcluded ? [excludedFilter(apply, parameters)] : []),
-        ...(buildOSFilterConfig
-            ? [
-                buildOSFilterConfig(
-                    {
-                        label: intl.formatMessage(messages.osFilterLabel),
-                        type: 'checkbox',
-                        id: 'rhel_version',
-                        value: toGroupSelectionValue(
-                            parameters.rhel_version
-                                ? parameters.rhel_version.split(',')
-                                : []
-                        ),
-                        onChange: (event, value) => {
-                            /* `versions` must be of type string, e.g., "8.9,9.0" */
-                            const versions = Object.values(value)
-                                .flatMap((versions) => Object.keys(versions))
-                                .toString();
-                            apply({ rhel_version: versions, page: 1 });
-                        }
-                    },
-                    operatingSystems
-                )
-            ]
-            : [])
+        ...osVersionFilter
     ];
 
     return <Fragment>
